requirements.txt
pandas
google-generativeai
streamlit
plotly
